requirements.txt
# Backend requirements.txt

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
pyjwt[crypto]==2.8.0
httpx==0.26.0

# AI and Document Processing
google-genai==0.3.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.2
pdf2image==1.16.3

# Database
supabase==2.10.0

# Utilities
requests==2.31.0
pydantic==2.7.4
pydantic-settings==2.3.4